Bugs fixed & simplify some codes

+++ simNeural/Neural_Algorithms.h
@@ -56,8 +56,8 @@ namespace neu_alg {
     }
     
     static void conv_descent_gradient(double lr, Eigen::MatrixXd& in1, Eigen::MatrixXd& in2, Eigen::MatrixXd& result, int kernel_size) {
-        int resultRow = result.rows(),
-        resultCol = result.cols();
+        int resultRow = static_cast<int>(result.rows()),
+        resultCol = static_cast<int>(result.cols());
         
         for (int r = 0; r < resultRow; r++) {
             for (int c = 0; c < resultCol; c++) {
